Added command acknowledge for prosthesis server
ACK messages are displayed in the client test.

Server should be echoing back commands to acknoledge their receipt which
the client can use to check and see if the server is hung up on a
command. There are no heartbeats yet however, the plan is to include
both a watchdog and heartbeat into the system to detect hangs.