Now that we're receiving real telemetry and not
just string/string data the data parsing needed to be changed to just
plucking out the ID, and passing the data on to the relevant handlers.

The motor controller arduino class now parses the output of the Arduino
without issue!

Changed the documentation of the motor controller telemetry so that it
shows up in intellisense nicely.