Prosthesis OS successfully connects to the Arduino
and processes/forwards telemetry data to clients.

The state machine is pretty hacked up right now since we're no longer
having clients able to control/send commands to prosthesis. Also, it
looks like I need to re-evaluate how state changes and faults work as
the arduino connection seems sketchy when reconnecting. This isn't
likely to happen in the Prosthesis environment but we should be able to
do it anyways.