Changes made to support heart beat signals from Arduinos
Added public API for enabling and disabling a heartbeat signal from the
Arduinos. The heartbeat period and number of allowable missed deadlines
are configurable through this API.

We should now be able to reliably detect when an Arduino is disconnected
(USB disconnect) or becomes unresponsive.

Added hooks to the disconnect event from within the ProsthesisOS state
machine. Added the sensor arduino to prosthesis OS.

Added a prosthesis logger to our unit tests so we can capture verbose
information about our unit tests when needed. Added a slew of additional
unit tests to detect different disconnection scenarios.