Adjusting baud rate of the arduinos to a higher speed
Added telemetry logging to a CSV in the arduino comms test so we can
gather telemetry during our testing.

Capped our telemetry rate to a max period of 50ms in the testing. This
should be the upper limit for the entire system otherwise we run into
the upper limit of serial transfer.

For telemetry periods below 50, the unit tests won't treat failure to
produce telemetry at that rate as a fail case as it is beyond the specs
for our system but it's good to know the limits. Added a 30 minute soak
test to ensure the system can reliably deliver telmetry for at least
this period of time. Ideally, it should be double our expected period of
operation.