Added ability to skip soak test for quick unit testing
Vastly improved the telemetry simulator: it now has sliders and numeric
up/down controls in addition to output mirroring for motor values.

Updated telemetry class to conform to the arduino telemetry including
both pressure setpoints. Added enum to the telemetry class to
differentiate which entry in the array belongs to which hydraulic
system.